LoaderBar moved as separate sub-component
